refactor(Hoster): restructured packages
GH-34: restructured packages and renamed ExcelFile to ExcelFileEntity
