+++ hoster/src/main/java/com/excel/dto/ExcelFileDTO.java
@@ -1,4 +1,4 @@
-package com.excel.hoster.excelfile;
+package com.excel.dto;
 
 import jakarta.validation.constraints.NotBlank;
