+++ hoster/src/main/java/com/excel/hoster/controller/ExposeExcelController.java
@@ -1,9 +1,10 @@
 package com.excel.hoster.controller;
 
 
-import com.excel.hoster.excelfile.ExcelFile;
-import com.excel.hoster.excelfile.ExcelFileService;
-import com.excel.hoster.excelfile.ExcelRepository;
+import com.excel.hoster.repository.entity.ExcelFileEntity;
+import com.excel.hoster.service.ExcelFileService;
+import com.excel.hoster.repository.ExcelRepository;
+import com.excel.hoster.service.ObjectResponse;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -47,21 +48,21 @@ public class ExposeExcelController {
     @GetMapping("/getExcelVersion")
     public ResponseEntity<?> getExcelVersion(@RequestParam(name="fullTextId",required = true) String fullTextId) {
 
-        ExcelFile excelFile = excelFileService.getExcelFileByFullTextId(fullTextId);
+        ExcelFileEntity excelFile = excelFileService.getExcelFileByFullTextId(fullTextId);
 
         if (excelFile ==null) {
             ObjectResponse<String> response = new ObjectResponse<>(HttpStatus.NOT_FOUND.value(), "No Excel found with fullTextId: " + fullTextId,null);
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
         }
 
-        ObjectResponse<ExcelFile> response = new ObjectResponse<>(HttpStatus.OK.value(), "Excel version successfully requested",excelFile);
+        ObjectResponse<ExcelFileEntity> response = new ObjectResponse<>(HttpStatus.OK.value(), "Excel version successfully requested",excelFile);
         return ResponseEntity.ok(response);
     }
 
     @GetMapping("/getExcelFile")
     public ResponseEntity<?> getExcelFile(@RequestParam(name = "fullTextId", required = true) String fullTextId) {
 
-        ExcelFile excelFile = excelFileService.getExcelFileByFullTextId(fullTextId);
+        ExcelFileEntity excelFile = excelFileService.getExcelFileByFullTextId(fullTextId);
 
         if (excelFile ==null) {
             ObjectResponse<String> response = new ObjectResponse<>(HttpStatus.NOT_FOUND.value(), "No Excel found with fullTextId: " + fullTextId,null);
