--- hoster/src/main/java/com/excel/hoster/excelfile/ExcelRepository.java
@@ -1,7 +1,0 @@
-package com.excel.hoster.excelfile;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface ExcelRepository extends CrudRepository<ExcelFile, String> {
-
-}
