+++ hoster/src/main/java/com/excel/hoster/repository/ExcelRepository.java
@@ -1,0 +1,8 @@
+package com.excel.hoster.repository;
+
+import com.excel.hoster.repository.entity.ExcelFileEntity;
+import org.springframework.data.repository.CrudRepository;
+
+public interface ExcelRepository extends CrudRepository<ExcelFileEntity, String> {
+
+}
