+++ hoster/src/main/resources/application.properties
@@ -9,5 +9,3 @@ spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 #Custom Properties
 apache.poi.version=4.1.2
-
-#this is roland testing out feature branches
